Clean up tests, move some test utils to a common module.
