+++ api/tests/python/end_points/auto_analyse/common.py
@@ -2,35 +2,9 @@ import pytest
 from datetime import date
 
 from namex.constants import EntityTypes
-from namex.models import User
 
 API_BASE_URI = '/api/v1/'
 ENDPOINT_PATH = API_BASE_URI + 'name-analysis'
-
-token_header = {
-    "alg": "RS256",
-    "typ": "JWT",
-    "kid": "flask-jwt-oidc-test-client"
-}
-
-claims = {
-    "iss": "https://sso-dev.pathfinder.gov.bc.ca/auth/realms/sbc",
-    "sub": "43e6a245-0bf7-4ccf-9bd0-e7fb85fd18cc",
-    "aud": "NameX-Dev",
-    "exp": 31531718745,
-    "iat": 1531718745,
-    "jti": "flask-jwt-oidc-test-support",
-    "typ": "Bearer",
-    "username": "test-user",
-    "realm_access": {
-        "roles": [
-            "{}".format(User.EDITOR),
-            "{}".format(User.APPROVER),
-            "viewer",
-            "user"
-        ]
-    }
-}
 
 
 @pytest.mark.skip
