+++ api/tests/python/end_points/auto_analyse/protected_names/test_end_designation_more_than_once.py
@@ -5,7 +5,7 @@ from urllib.parse import quote_plus
 
 from namex.services.name_request.auto_analyse import AnalysisIssueCodes
 
-from ...common import assert_issues_count_is_gt, assert_has_word_upper, assert_has_designations_upper, \
+from ..common import assert_issues_count_is_gt, assert_has_word_upper, assert_has_designations_upper, \
     save_words_list_classification, assert_has_issue_type, assert_has_no_issue_type
 from ..common import ENDPOINT_PATH
 from ...common import token_header, claims
