+++ api/tests/python/end_points/auto_analyse/protected_names/test_name_requires_consent_more_than_one_word_request.py
@@ -5,7 +5,7 @@ from urllib.parse import quote_plus
 
 from namex.services.name_request.auto_analyse import AnalysisIssueCodes
 
-from ...common import assert_issues_count_is_gt, assert_has_word_upper, save_words_list_classification, save_words_list_virtual_word_condition
+from ..common import assert_issues_count_is_gt, assert_has_word_upper, save_words_list_classification, save_words_list_virtual_word_condition
 from ..common import ENDPOINT_PATH
 from ...common import token_header, claims
