+++ api/tests/python/end_points/auto_analyse/protected_names/test_request_actions_bc.py
@@ -5,7 +5,7 @@ from urllib.parse import quote_plus
 
 from ..common import save_words_list_classification, assert_issues_count_is_zero
 from ..common import ENDPOINT_PATH
-from ..common import token_header, claims
+from ...common import token_header, claims
 
 
 @pytest.mark.xfail(raises=ValueError)
