+++ api/tests/python/end_points/auto_analyse/protected_names/test_too_many_words_request.py
@@ -5,7 +5,7 @@ from urllib.parse import quote_plus
 
 from namex.services.name_request.auto_analyse import AnalysisIssueCodes
 
-from ...common import assert_issues_count_is_gt, assert_issue_type_is_one_of, assert_has_issue_type, save_words_list_classification
+from ..common import assert_issues_count_is_gt, assert_issue_type_is_one_of, assert_has_issue_type, save_words_list_classification
 from ..common import ENDPOINT_PATH
 from ...common import token_header, claims
