+++ api/tests/python/end_points/common/README.md
@@ -1,0 +1,6 @@
+#### Move all common test utils related to endpoints here
+
+- configuration
+- auth
+- http
+- logging
