+++ api/tests/python/end_points/common/__init__.py
@@ -1,0 +1,1 @@
+from .configuration import *
