+++ api/tests/python/end_points/common/logging.py
@@ -1,0 +1,2 @@
+def log_request_path(path):
+    print('\n' + 'request: ' + path + '\n')
