+++ api/tests/python/end_points/payments/test_payments.py
@@ -1,9 +1,10 @@
 import json
 
-from urllib.parse import quote_plus
-
 from .common import API_BASE_URI
-from ..common import token_header, claims
+# Import token and claims if you need it
+# from ..common import token_header, claims
+from ..common.http import build_test_query, build_request_uri
+from ..common.logging import log_request_path
 
 create_payment_request = {
     "payment_info": {
@@ -43,36 +44,15 @@ calculate_fees_request = {
 }
 
 
-def build_test_query(test_params):
-    query = ''
-
-    for entry in test_params:
-        query += '&'.join("{!s}={}".format(k, quote_plus(v)) for (k, v) in entry.items())
-
-    return query
-
-
-def log_request_path(path):
-    print('\n' + 'request: ' + path + '\n')
-
-
-def setup_test_token(client, jwt, app):
-    # Create JWT & setup header with a Bearer Token using the JWT
-    token = jwt.create_jwt(claims, token_header)
-    headers = {'Authorization': 'Bearer ' + token, 'content-type': 'application/json'}
-
-    return token, headers
-
-
 def test_get_payment(client, jwt, app):
     payment_id = 'abcd153'
     request_uri = API_BASE_URI + payment_id
     test_params = [{}]
 
     query = build_test_query(test_params)
-
-    path = request_uri + '?' + query if len(query) > 0 else request_uri
+    path = build_request_uri(request_uri, query)
     log_request_path(path)
+
     response = client.get(path)
     payload = json.loads(response.data)
 
@@ -84,8 +64,9 @@ def test_create_payment(client, jwt, app):
     request_uri = API_BASE_URI
 
     path = request_uri
-    log_request_path(path)
     body = json.dumps(create_payment_request)
+    log_request_path(path)
+
     response = client.post(path, json=body)
     payload = json.loads(response.data)
 
@@ -100,8 +81,9 @@ def test_update_payment(client, jwt, app):
     request_uri = API_BASE_URI + payment_id
 
     path = request_uri
-    log_request_path(path)
     body = json.dumps(create_payment_request)
+    log_request_path(path)
+
     response = client.put(path, json=body)
     payload = json.loads(response.data)
 
@@ -119,9 +101,9 @@ def test_get_invoice(client, jwt, app):
     test_params = [{'invoice_id': invoice_id}]
 
     query = build_test_query(test_params)
-
-    path = request_uri + '?' + query if len(query) > 0 else request_uri
+    path = build_request_uri(request_uri, query)
     log_request_path(path)
+
     response = client.get(path)
 
     assert response.status_code == 200
@@ -137,9 +119,9 @@ def test_get_invoices(client, jwt, app):
     test_params = [{}]
 
     query = build_test_query(test_params)
-
-    path = request_uri + '?' + query if len(query) > 0 else request_uri
+    path = build_request_uri(request_uri, query)
     log_request_path(path)
+
     response = client.get(path)
 
     assert response.status_code == 200
@@ -157,9 +139,9 @@ def test_get_transaction(client, jwt, app):
     }]
 
     query = build_test_query(test_params)
-
-    path = request_uri + '?' + query if len(query) > 0 else request_uri
+    path = build_request_uri(request_uri, query)
     log_request_path(path)
+
     response = client.get(path)
 
     assert response.status_code == 200
@@ -174,9 +156,9 @@ def test_get_transactions(client, jwt, app):
     test_params = [{}]
 
     query = build_test_query(test_params)
-
-    path = request_uri + '?' + query if len(query) > 0 else request_uri
+    path = build_request_uri(request_uri, query)
     log_request_path(path)
+
     response = client.get(path)
 
     assert response.status_code == 200
@@ -195,9 +177,9 @@ def test_create_transaction(client, jwt, app):
     }]
 
     query = build_test_query(test_params)
-
-    path = request_uri + '?' + query if len(query) > 0 else request_uri
+    path = build_request_uri(request_uri, query)
     log_request_path(path)
+
     response = client.post(path)
     payload = json.loads(response.data)
 
@@ -213,13 +195,13 @@ def test_update_transaction(client, jwt, app):
     test_params = [{}]
 
     query = build_test_query(test_params)
-
-    path = request_uri + '?' + query if len(query) > 0 else request_uri
-    log_request_path(path)
+    path = build_request_uri(request_uri, query)
     body = json.dumps({
         "receipt_number": receipt_number,
         "transaction_identifier": transaction_id
     })
+    log_request_path(path)
+
     response = client.put(path, json=body)
     payload = json.loads(response.data)
 
@@ -232,9 +214,9 @@ def test_get_receipt(client, jwt, app):
     test_params = [{}]
 
     query = build_test_query(test_params)
-
-    path = request_uri + '?' + query if len(query) > 0 else request_uri
+    path = build_request_uri(request_uri, query)
     log_request_path(path)
+
     response = client.get(path)
 
     assert response.status_code == 200
@@ -252,8 +234,9 @@ def test_calculate_fees(client, jwt, app):
     request_uri = API_BASE_URI + 'fees'
 
     path = request_uri
-    log_request_path(path)
     body = json.dumps(calculate_fees_request)
+    log_request_path(path)
+
     response = client.post(path, json=body)
     payload = json.loads(response.data)
